+++ run/kubejs/startup_scripts/block/crops.js
@@ -1,7 +1,9 @@
 StartupEvents.registry('block', e => {
-    e.create('crop', 'tfc:crop');
+    e.create('crop', 'tfc:crop')
+        .model('minecraft:block/deepslate');
     e.create('flooded_crop', 'tfc:flooded_crop');
-    e.create('pickable_crop', 'tfc:pickable_crop');
+    e.create('pickable_crop', 'tfc:pickable_crop')
+        .setModel(m => m.parent('minecraft:block/light_blue_wool'));
     e.create('spreading_crop', 'tfc:spreading_crop')
         .deadBlock(d => {
             d.models((stage, m) => {
@@ -35,4 +37,28 @@ StartupEvents.registry('block', e => {
         .seeds('kubejs:crop_seeds')
         .food('minecraft:chorus_fruit')
         .deadModel('minecraft:block/cobblestone');
+
+    e.create('crop_2', 'tfc:crop')
+        .expiryModifier(250)
+        .stages(2)
+        .model(0, 'minecraft:block/white_wool')
+        .setModel(1, m => m.parent('minecraft:block/red_wool'));
+    e.create('crop_4', 'tfc:crop')
+        .growthModifier(2)
+        .stages(4)
+        .model(0, 'minecraft:block/white_wool')
+        .model(1, 'minecraft:block/red_wool')
+        .model(2, 'minecraft:block/pink_wool')
+        .model(3, 'minecraft:block/black_wool');
+    e.create('crop_8', 'tfc:crop')
+        .growthModifier(4)
+        .stages(8)
+        .model(0, 'minecraft:block/white_wool')
+        .model(1, 'minecraft:block/red_wool')
+        .model(2, 'minecraft:block/pink_wool')
+        .model(3, 'minecraft:block/black_wool')
+        .model(4, 'minecraft:block/green_wool')
+        .model(5, 'minecraft:block/yellow_wool')
+        .model(6, 'minecraft:block/lime_wool')
+        .model(7, 'minecraft:block/blue_wool')
 })