+++ run/kubejs/startup_scripts/events/food_trait.js
@@ -1,0 +1,3 @@
+TFCEvents.registerFoodTrait(e => {
+    e.registerTraitWithTooltip(0.01, 'kubejs:sealed');
+})