+++ src/main/java/com/notenoughmail/kubejs_tfc/block/DefaultCropBlockBuilder.java
@@ -14,6 +14,6 @@ public class DefaultCropBlockBuilder extends AbstractCropBlockBuilder {
 
     @Override
     public Block createObject() {
-        return CropUtils.defaultCrop(createExtendedProperties(), stages, dead, seeds, nutrient, climateRange);
+        return CropUtils.defaultCrop(createExtendedProperties(), stages, dead, seeds, nutrient, climateRange, growthMod, expiryMod);
     }
 }