+++ src/main/java/com/notenoughmail/kubejs_tfc/util/implementation/CropUtils.java
@@ -1,5 +1,6 @@
 package com.notenoughmail.kubejs_tfc.util.implementation;
 
+import com.notenoughmail.kubejs_tfc.util.implementation.custom.block.ICustomCropBlock;
 import net.dries007.tfc.common.blockentities.FarmlandBlockEntity;
 import net.dries007.tfc.common.blocks.ExtendedProperties;
 import net.dries007.tfc.common.blocks.TFCBlockStateProperties;
@@ -20,10 +21,20 @@ public class CropUtils {
             Supplier<? extends Block> dead,
             Supplier<? extends Item> seeds,
             FarmlandBlockEntity.NutrientType primaryNutrient,
-            Supplier<ClimateRange> climateRange
+            Supplier<ClimateRange> climateRange,
+            Supplier<Double> growth,
+            Supplier<Double> expiry
     ) {
         final IntegerProperty property = TFCBlockStateProperties.getAgeProperty(stages);
-        return new DefaultCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange) {
+        return new ExtCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange) {
+            @Override
+            public float growthModifier() {
+                return growth.get().floatValue();
+            }
+            @Override
+            public float expiryModifier() {
+                return expiry.get().floatValue();
+            }
             @Override
             public IntegerProperty getAgeProperty() {
                 return property;
@@ -37,10 +48,20 @@ public class CropUtils {
             Supplier<? extends Block> dead,
             Supplier<? extends Item> seeds,
             FarmlandBlockEntity.NutrientType primaryNutrient,
-            Supplier<ClimateRange> climateRange
+            Supplier<ClimateRange> climateRange,
+            Supplier<Double> growth,
+            Supplier<Double> expiry
     ) {
         final IntegerProperty property = TFCBlockStateProperties.getAgeProperty(stages);
-        return new FloodedCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange) {
+        return new ExtFloodedCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange) {
+            @Override
+            public float growthModifier() {
+                return growth.get().floatValue();
+            }
+            @Override
+            public float expiryModifier() {
+                return expiry.get().floatValue();
+            }
             @Override
             public IntegerProperty getAgeProperty() {
                 return property;
@@ -56,10 +77,20 @@ public class CropUtils {
             FarmlandBlockEntity.NutrientType primaryNutrient,
             Supplier<ClimateRange> climateRange,
             @Nullable Supplier<Supplier<? extends Item>> fruit,
-            Supplier<Supplier<? extends Item>> matureFruit
+            Supplier<Supplier<? extends Item>> matureFruit,
+            Supplier<Double> growth,
+            Supplier<Double> expiry
     ) {
         final IntegerProperty property = TFCBlockStateProperties.getAgeProperty(stages);
-        return new PickableCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange, fruit, matureFruit) {
+        return new ExtPickableCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange, fruit, matureFruit) {
+            @Override
+            public float growthModifier() {
+                return growth.get().floatValue();
+            }
+            @Override
+            public float expiryModifier() {
+                return expiry.get().floatValue();
+            }
             @Override
             public IntegerProperty getAgeProperty() {
                 return property;
@@ -74,11 +105,21 @@ public class CropUtils {
             Supplier<? extends Item> seeds,
             FarmlandBlockEntity.NutrientType primaryNutrient,
             Supplier<ClimateRange> climateRange,
-            Supplier<Supplier<? extends Block>> fruitBlock
+            Supplier<Supplier<? extends Block>> fruitBlock,
+            Supplier<Double> growth,
+            Supplier<Double> expiry
 
     ) {
         final IntegerProperty property = TFCBlockStateProperties.getAgeProperty(stages);
-        return new SpreadingCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange, fruitBlock) {
+        return new ExtSpreadingCropBlock(properties, stages, dead, seeds, primaryNutrient, climateRange, fruitBlock) {
+            @Override
+            public float growthModifier() {
+                return growth.get().floatValue();
+            }
+            @Override
+            public float expiryModifier() {
+                return expiry.get().floatValue();
+            }
             @Override
             public IntegerProperty getAgeProperty() {
                 return property;
@@ -94,22 +135,76 @@ public class CropUtils {
             Supplier<? extends Item> seeds,
             FarmlandBlockEntity.NutrientType primaryNutrient,
             Supplier<ClimateRange> climateRange,
-            boolean requiresStick
+            boolean requiresStick,
+            Supplier<Double> growth,
+            Supplier<Double> expiry
     ) {
         final IntegerProperty property = TFCBlockStateProperties.getAgeProperty(singleStages + doubleStages);
         if (requiresStick) {
-            return new ClimbingCropBlock(properties, singleStages, singleStages + doubleStages, dead, seeds, primaryNutrient, climateRange) {
+            return new ExtClimbingCropBlock(properties, singleStages, singleStages + doubleStages, dead, seeds, primaryNutrient, climateRange) {
+                @Override
+                public float growthModifier() {
+                    return growth.get().floatValue();
+                }
+                @Override
+                public float expiryModifier() {
+                    return expiry.get().floatValue();
+                }
                 @Override
                 public IntegerProperty getAgeProperty() {
                     return property;
                 }
             };
         }
-        return new DoubleCropBlock(properties, singleStages - 1, singleStages + doubleStages - 1, dead, seeds, primaryNutrient, climateRange) {
+        return new ExtDoubleCropBlock(properties, singleStages - 1, singleStages + doubleStages - 1, dead, seeds, primaryNutrient, climateRange) {
+            @Override
+            public float growthModifier() {
+                return growth.get().floatValue();
+            }
+            @Override
+            public float expiryModifier() {
+                return expiry.get().floatValue();
+            }
             @Override
             public IntegerProperty getAgeProperty() {
                 return property;
             }
         };
+    }
+
+    private static abstract class ExtCropBlock extends DefaultCropBlock implements ICustomCropBlock {
+        protected ExtCropBlock(ExtendedProperties properties, int maxAge, Supplier<? extends Block> dead, Supplier<? extends Item> seeds, FarmlandBlockEntity.NutrientType primaryNutrient, Supplier<ClimateRange> climateRange) {
+            super(properties, maxAge, dead, seeds, primaryNutrient, climateRange);
+        }
+    }
+
+    private static abstract class ExtFloodedCropBlock extends FloodedCropBlock implements ICustomCropBlock {
+        protected ExtFloodedCropBlock(ExtendedProperties properties, int maxAge, Supplier<? extends Block> dead, Supplier<? extends Item> seeds, FarmlandBlockEntity.NutrientType primaryNutrient, Supplier<ClimateRange> climateRange) {
+            super(properties, maxAge, dead, seeds, primaryNutrient, climateRange);
+        }
+    }
+
+    private static abstract class ExtPickableCropBlock extends PickableCropBlock implements ICustomCropBlock {
+        protected ExtPickableCropBlock(ExtendedProperties properties, int maxAge, Supplier<? extends Block> dead, Supplier<? extends Item> seeds, FarmlandBlockEntity.NutrientType primaryNutrient, Supplier<ClimateRange> climateRange, @Nullable Supplier<Supplier<? extends Item>> fruit, Supplier<Supplier<? extends Item>> matureFruit) {
+            super(properties, maxAge, dead, seeds, primaryNutrient, climateRange, fruit, matureFruit);
+        }
+    }
+
+    private static abstract class ExtSpreadingCropBlock extends SpreadingCropBlock implements ICustomCropBlock {
+        protected ExtSpreadingCropBlock(ExtendedProperties properties, int maxAge, Supplier<? extends Block> dead, Supplier<? extends Item> seeds, FarmlandBlockEntity.NutrientType primaryNutrient, Supplier<ClimateRange> climateRange, Supplier<Supplier<? extends Block>> fruit) {
+            super(properties, maxAge, dead, seeds, primaryNutrient, climateRange, fruit);
+        }
+    }
+
+    private static abstract class ExtClimbingCropBlock extends ClimbingCropBlock implements ICustomCropBlock {
+        protected ExtClimbingCropBlock(ExtendedProperties properties, int maxSingleAge, int maxAge, Supplier<? extends Block> dead, Supplier<? extends Item> seeds, FarmlandBlockEntity.NutrientType primaryNutrient, Supplier<ClimateRange> climateRange) {
+            super(properties, maxSingleAge, maxAge, dead, seeds, primaryNutrient, climateRange);
+        }
+    }
+
+    private static abstract class ExtDoubleCropBlock extends DoubleCropBlock implements ICustomCropBlock {
+        protected ExtDoubleCropBlock(ExtendedProperties properties, int maxSingleAge, int maxAge, Supplier<? extends Block> dead, Supplier<? extends Item> seeds, FarmlandBlockEntity.NutrientType primaryNutrient, Supplier<ClimateRange> climateRange) {
+            super(properties, maxSingleAge, maxAge, dead, seeds, primaryNutrient, climateRange);
+        }
     }
 }