+++ src/main/java/com/notenoughmail/kubejs_tfc/util/implementation/custom/block/ICustomCropBlock.java
@@ -1,0 +1,8 @@
+package com.notenoughmail.kubejs_tfc.util.implementation.custom.block;
+
+public interface ICustomCropBlock {
+
+    float growthModifier();
+
+    float expiryModifier();
+}