+++ changelog.md
@@ -5,7 +5,10 @@
 - Fix double crops not having yield multipliers in their loot tables
 - Add convenience method for setting the dead model of torches
 - Add sapling block builder
-- Rename `#productItem(ResourceLocation)` to `#existingProductItem(ResourceLocation)` in crop block builders to prevent Rhino throwing a fit
+- Rename several methods in crop block block builders to prevent Rhino throwing a fit when used
+  - `#productItem(ResourceLocation)` -> `#existingProductItem(ResourceLocation)`
+  - `#model(number,Consumer)` -> `#setModel(number,Consumer)`
+  - `#model(Consumer)` -> `#setModel(Consumer)`
 - Allow `RecipeJS#getOriginalRecipeIngredients()` to work with TFC recipe types
 
 ### 1.3.1